KPV-2334:: Added button for contest application report parallel to participatory budget propousal report
