+++ grails-app/controllers/kuorum/contest/ContestController.groovy
@@ -9,7 +9,6 @@ import kuorum.util.TimeZoneUtil
 import kuorum.web.commands.payment.CampaignContentCommand
 import kuorum.web.commands.payment.CampaignSettingsCommand
 import kuorum.web.commands.payment.contest.*
-
 import kuorum.web.constants.WebConstants
 import org.kuorum.rest.model.communication.contest.*
 import org.kuorum.rest.model.kuorumUser.BasicDataKuorumUserRSDTO
@@ -17,6 +16,8 @@ import org.kuorum.rest.model.notification.campaign.CampaignStatusRSDTO
 import org.kuorum.rest.model.search.DirectionDTO
 
 import java.lang.reflect.UndeclaredThrowableException
+
+import static kuorum.util.rest.RestKuorumApiService.ApiMethod.ACCOUNT_CONTEST_REPORT
 
 class ContestController extends CampaignController {
 
@@ -302,7 +303,7 @@ class ContestController extends CampaignController {
                         activityType  : contestApplicationRSDTO.activityType,
                         focusType     : contestApplicationRSDTO.focusType,
                         url           : g.createLink(mapping: 'contestApplicationShow', params: contestApplicationRSDTO.encodeAsLinkProperties()),
-                        numVotes         : contestApplicationRSDTO.votes
+                        numVotes      : contestApplicationRSDTO.votes
                 ]
             }
         }
@@ -366,7 +367,7 @@ class ContestController extends CampaignController {
     }
 
     def sendApplicationsReport() {
-        render "OK";
+        abstractSendProposalsReport(ACCOUNT_CONTEST_REPORT)
     }
 
 //    @Secured(['ROLE_CAMPAIGN_CONTEST_APPLICATION'])
