+++ grails-app/services/kuorum/util/rest/RestKuorumApiService.groovy
@@ -297,6 +297,7 @@ class RestKuorumApiService {
         ACCOUNT_CONTEST("/communication/campaign/contest/{userId}/{campaignId}"),
         ACCOUNT_CONTEST_COPY("/communication/campaign/contest/{userId}/{campaignId}/copy"),
         ACCOUNT_CONTEST_APPLICATIONS_LIST("/communication/campaign/contest/{userId}/{campaignId}/applications"),
+        ACCOUNT_CONTEST_REPORT("/communication/campaign/contest/{userId}/{campaignId}/report"),
 
         ACCOUNT_CONTEST_APPLICATIONS("/communication/campaign/contest/application/{userId}"),
         ACCOUNT_CONTEST_APPLICATION("/communication/campaign/contest/application/{userId}/{campaignId}"),
